configured embedded youtube player
Used the youtube api options to remove title and share bar, remove the
youtube logo, and enable javascript controls. Also set a default width
and height. These don't serve a purpose in terms of window size, but the
idea was to encourage the player to not load the 360p version of the
video.